体重值精度问题

+++ app/src/main/java/com/example/han/referralproject/health/model/DetectResult.java
@@ -36,7 +36,7 @@ public class DetectResult implements Serializable {
     public String watchState;
     public String hypoglycemia;
     public String stapleFood;
-    public int weight;
+    public float weight;
     public String wineDrink;
     public String currentPhoto;//体检报告的头像   key  "detectHeadIcon"
     public String psychologicalRecovery;//心理调整 1良好 2一般 3差 ,
@@ -174,7 +174,7 @@ public class DetectResult implements Serializable {
         this.watchState = watchState;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(float weight) {
         this.weight = weight;
     }
